refactor(Timeline): Use `TimelineGrid`
